[render] DrawableTrackBall and Axis moved in TrackBallViewerDrawer
